more field tests
